refactor(agent): extract event forwarder from manager.rs
`spawn_v2_event_forwarder` plus its three helpers (`summarize_input`,
`flush_thinking`, `flush_text`) move out of `src/agent/manager.rs`
(1066 lines) into a dedicated `src/agent/event_forwarder.rs` (397
lines including docs). Pure relocation — no behavior change, no API
change outside the `agent` module.

Why: manager.rs was hosting two distinct concepts side-by-side.

1. Agent lifecycle (what HTTP handlers call) — `start_agent`,
   `stop_agent`, `sleep_agent`, `notify_agent`, `stop_all`,
   `get_running_agent_names`, `resolve_bridge_endpoint`, the
   `AgentLifecycle` trait impl. Owns the `agents` map, the driver
   registry, and the mutex. ~400 lines of legitimate orchestration.

2. Driver-event translation (what runs in the background) — a 270-line
   free function reading `mpsc::Receiver<DriverEvent>` and fanning
   out to `activity_logs`, `trace_store`, the `trace_tx` broadcast,
   and `store.agent_sessions`. Six `DriverEvent` variants, each with
   its own buffering/flushing logic. Already loosely coupled: took
   seven owned / `Arc` arguments and only touched manager state
   (agents map) in one spot (deferred-notification delivery in the
   `Completed` arm).

The forwarder reads like its own concept; it wanted its own file.

Changes:
- New `src/agent/event_forwarder.rs` — public-in-module
  `spawn_v2_event_forwarder` + private helpers.
- `V2Agent` struct and its three fields promoted from private to
  `pub(super)` so the sibling forwarder module can access
  `handle`, `_event_tasks`, and `pending_notification_count`.
  Visibility is scoped to the `agent` module only.
- Call site in `AgentManager::start_agent` updated to
  `super::event_forwarder::spawn_v2_event_forwarder(...)`.
- Dead imports purged from `manager.rs` (`tokio::sync::broadcast`,
  `tracing::{error, trace}`, `AgentEventItem`, `DriverEvent`,
  `TraceEvent`).

Payoff:
- manager.rs: 1066 → 730 lines. Concept matches the filename:
  lifecycle + trait impl, nothing else.
- event_forwarder.rs: already structured for in-isolation testing.
  Feeding it a scripted `Receiver<DriverEvent>` would cover the
  six variants + flush ordering + deferred-notification delivery
  without needing a live runtime binary. No tests added in this
  PR — that's a separate scope.

Verification: `cargo test` 362 pass (no change), `cargo clippy
-D warnings` clean, `cargo fmt` clean.
